ReInitial Commit

+++ src/main/java/com/example/demoProjectoLabBack/model/dto/WorkerProfileDto.java
@@ -7,12 +7,12 @@ public class WorkerProfileDto {
     private String description;
     private int dni;
     private String direccion;
-    private int rating;
+    private double rating;
     private UserDto user;
     private List<String> jobTitles;
 
     // Constructor
-    public WorkerProfileDto(String id, String description, int dni, String direccion, int rating, UserDto user, List<String> jobTitles) {
+    public WorkerProfileDto(String id, String description, int dni, String direccion, double rating, UserDto user, List<String> jobTitles) {
         this.id = id;
         this.description = description;
         this.dni = dni;
@@ -31,7 +31,7 @@ public class WorkerProfileDto {
     public void setDni(int dni) { this.dni = dni; }
     public String getDireccion() { return direccion; }
     public void setDireccion(String direccion) { this.direccion = direccion; }
-    public int getRating() { return rating; }
+    public double getRating() { return rating; }
     public void setRating(int rating) { this.rating = rating; }
     public UserDto getUser() { return user; }
     public void setUser(UserDto user) { this.user = user; }
